+++ scout/server/blueprints/variants/views.py
@@ -533,8 +533,13 @@ def upload_panel(institute_id, case_name):
 @variants_bp.route("/verified", methods=["GET"])
 def download_verified():
     """Download all verified variants for user's cases"""
-    user_obj = store.user(current_user.email)
-    user_institutes = user_obj.get("institutes")
+
+    flash(current_user.__dict__)
+    user_institutes = (
+        [inst["_id"] for inst in store.institutes()]
+        if current_user.is_admin
+        else current_user.get("institutes")
+    )
     temp_excel_dir = os.path.join(variants_bp.static_folder, "verified_folder")
     os.makedirs(temp_excel_dir, exist_ok=True)